fix copy pasta bug in AsyncFd
this was preventing every ping from succeeding -- instead of waiting for
the socket to become writable when sending icmp echo requests, we were
accidentally (due to copy pasta) waiting for it to become readable.
because the socket is only listening for ICMP packets, this never
happens since we never send a request in the first place.