fix memory leak in MaybeUninit<u8> handling
Prior to this commit, when receiving bytes from a socket we would only
transmute written `MaybeUninit<u8>` to if and only if the number of
bytes read were within a specific range. If the number of bytes read
fell outside that range we would simply continue the loop, assuming that
we hadn't received a valid ICMP packet. Continuing the loop would cause
the `[MaybeUninit<u8>]` to be dropped, which is problematic because
dropping a `MaybeUninit<T>` does nothing -- it doesn't even drop the
inner `T`, hence the memory leak (i think).